Move User models to rmndin.users.models
