+++ rmndin/rmndin/lib/db/enums.py
@@ -1,0 +1,6 @@
+import enum
+
+
+class DeliveryMethodEnum(enum.Enum):
+    reddit = "reddit"
+    email = "email"
