+++ rmndin/rmndin/lib/db/mixins.py
@@ -1,0 +1,27 @@
+import datetime
+import enum
+
+from rmndin import db
+
+
+class BaseMixin(object):
+    created = db.Column(db.DateTime, default=datetime.datetime.utcnow)
+    updated = db.Column(db.DateTime, default=datetime.datetime.utcnow,
+                        onupdate=datetime.datetime.utcnow)
+
+    def to_dict(self, exclude=None):
+        exclude = exclude or []
+
+        if isinstance(exclude, str):
+            exclude = [exclude]
+
+        rv = dict()
+        for column in self.__table__.columns:
+            name = column.name
+            if not name.startswith("_") and name not in exclude:
+                value = getattr(self, name)
+                if isinstance(value, enum.Enum):
+                    value = value.name
+
+                rv[name] = str(value)
+        return rv
